1) Исправлены ошибки в репозиториях 2) настройки для БД возвращены к веткам мастер/девелопмент 3) временно удалены скрипты миграции БД

+++ src/main/java/ru/auheal/repositories/AccentRepository.java
@@ -5,20 +5,9 @@ import org.springframework.stereotype.Repository;
 import ru.auheal.entities.Accent;
 
 /**
- * Репозиторий акцентов тренировок
+ * Репозиторий Акцентов тренировок
  */
-
 @Repository
-public interface AccentRepository extends JpaRepository<Accent, Long>{
-    /*
-     * СОГЛАШЕНИЕ О НАИМЕНОВАНИИ МЕТОДОВ РЕПОЗИТОРИЕВ
-     * Optional<Object> findById(Long id) найти объект по параметру
-     * List<Object> findAll() найти все объекты
-     * List<Object> findAllByEnabled(boolean enabled) найти все объекты по параметру
-     * void delete(Object object) удалить конкретный объект
-     * void deleteById(Long id) удалить объект по параметру
-     * void deleteAll(List<Object> objects) удалить список объектов
-     * Object save(Object object) сохранить объект
-     * List<Object> saveAll(List<Object> objects) сохранить список объектов
-     */
+public interface AccentRepository extends JpaRepository<Accent, Short>{
+
 }
