+++ src/main/java/ru/auheal/repositories/AuthorityRepository.java
@@ -5,22 +5,10 @@ import org.springframework.stereotype.Repository;
 import ru.auheal.entities.Authority;
 
 /**
- * Репозиторий ролей пользователей
+ * Репозиторий Ролей пользователей
  */
 @Repository
-public interface AuthorityRepository extends JpaRepository<Authority, Integer> {
-
-    /*
-     * СОГЛАШЕНИЕ О НАИМЕНОВАНИИ МЕТОДОВ РЕПОЗИТОРИЕВ
-     * Optional<Object> findById(Long id) найти объект по параметру
-     * List<Object> findAll() найти все объекты
-     * List<Object> findAllByEnabled(boolean enabled) найти все объекты по параметру
-     * void delete(Object object) удалить конкретный объект
-     * void deleteById(Long id) удалить объект по параметру
-     * void deleteAll(List<Object> objects) удалить список объектов
-     * Object save(Object object) сохранить объект
-     * List<Object> saveAll(List<Object> objects) сохранить список объектов
-     */
+public interface AuthorityRepository extends JpaRepository<Authority, Short> {
 
     /**
      * Найти роль пользователя по ее имени
