+++ src/main/java/ru/auheal/repositories/SportTypeRepository.java
@@ -5,20 +5,9 @@ import org.springframework.stereotype.Repository;
 import ru.auheal.entities.SportType;
 
 /**
- * Репозиторий видов спорта
+ * Репозиторий Видов спорта
  */
-
 @Repository
-public interface SportTypeRepository extends JpaRepository<SportType, Long> {
-    /*
-     * СОГЛАШЕНИЕ О НАИМЕНОВАНИИ МЕТОДОВ РЕПОЗИТОРИЕВ
-     * Optional<Object> findById(Long id) найти объект по параметру
-     * List<Object> findAll() найти все объекты
-     * List<Object> findAllByEnabled(boolean enabled) найти все объекты по параметру
-     * void delete(Object object) удалить конкретный объект
-     * void deleteById(Long id) удалить объект по параметру
-     * void deleteAll(List<Object> objects) удалить список объектов
-     * Object save(Object object) сохранить объект
-     * List<Object> saveAll(List<Object> objects) сохранить список объектов
-     */
+public interface SportTypeRepository extends JpaRepository<SportType, Short> {
+
 }
